chore: refactor indexer components into their own package
[ghstack-poisoned]
