+++ py_wtf/indexer/__init__.py
@@ -1,0 +1,4 @@
+from .file import index_dir, index_file
+from .pypi import index_project
+
+__all__ = ["index_file", "index_dir", "index_project"]
